requirements.txt
asgiref==3.3.4
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cssselect2==0.4.1
cycler==0.10.0
defusedxml==0.7.1
Django==3.2
django-bootstrap-form==3.4
django-durationwidget==1.0.5
django-filter==2.4.0
django-widget-tweaks==1.4.8
docopt==0.6.2
html5lib==1.1
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.1
mplcyberpunk==0.1.11
numpy==1.20.2
Pillow==8.2.0
pycparser==2.20
pyparsing==2.4.7
Pyphen==0.10.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
urllib3==1.26.4
WeasyPrint==52.5
webencodings==0.5.1
yarg==0.1.9
